修复用户主页正常索引
