+++ app/src/user/components/FeedbackList/index.css
@@ -1,21 +1,13 @@
-.feedback-title {
-  margin: 20px;
-  font-size: 16px;
+.feedback-list-item {
+  margin: 10px 0;
+  padding: 24px;
+  border: 1px solid #f0f0f0;
 }
 
-.feedback-item {
-  margin: 10px 0
+.icon-like:hover, .icon-like .done {
+  color: #1890ff;
 }
 
-.margin-t-10 {
-  margin-top: 10px;
-}
-
-.feedback-btn {
-  display: block;
-  margin: 20px auto 0;
-}
-
-.text-center {
-  text-align: center;
+.icon-dislike:hover, .icon-dislike .done {
+  color: #eb2f96;
 }
