--- app/src/user/components/FeedbackList2/index.css
@@ -1,13 +1,0 @@
-.feedback-list-item {
-  margin: 10px 0;
-  padding: 24px;
-  border: 1px solid #f0f0f0;
-}
-
-.icon-like:hover, .icon-like .done {
-  color: #1890ff;
-}
-
-.icon-dislike:hover, .icon-dislike .done {
-  color: #eb2f96;
-}
