+++ app/src/user/components/OldFeedbackList/index.css
@@ -1,0 +1,21 @@
+.feedback-title {
+  margin: 20px;
+  font-size: 16px;
+}
+
+.feedback-item {
+  margin: 10px 0
+}
+
+.margin-10 {
+  margin: 10px 0;
+}
+
+.feedback-btn {
+  display: block;
+  margin: 20px auto 0;
+}
+
+.text-center {
+  text-align: center;
+}
