+++ app/src/user/components/UserStatistic/index.css
@@ -1,0 +1,3 @@
+.text-center {
+  text-align: center;
+}
