+++ app/src/user/pages/Home/index.css
@@ -1,0 +1,13 @@
+.feedback-title {
+  margin: 20px;
+  font-size: 16px;
+}
+
+.margin-10 {
+  margin: 10px 0;
+}
+
+.feedback-btn {
+  display: block;
+  margin: 20px auto 0;
+}
